device: carry LH2 homographies as float32
The calibration message, the config page and the shared struct hold
each matrix element as IEEE-754 float32 in millimetres instead of an
int32 scaled by 1e3, and dotbot-libs moves with them. The config page
takes a new magic, so a page written in the old layout reads as
unprovisioned and the default net id applies until it is rewritten.

AI-assisted: Claude Opus 5

+++ device/bootloader/Source/localization.c
@@ -30,7 +30,7 @@ void localization_start(void) {
     _lh2_started = true;
 }
 
-void localization_init(int32_t homographies[][3][3], uint32_t homography_count) {
+void localization_init(float homographies[][3][3], uint32_t homography_count) {
     printf("Initialize localization with %u homography matrices\n", homography_count);
     localization_start();
 
@@ -38,7 +38,7 @@ void localization_init(int32_t homographies[][3][3], uint32_t homography_count) 
         printf("Store homography matrix for LH%u:\n", lh_index);
         for (int i = 0; i < 3; i++) {
             for (int j = 0; j < 3; j++) {
-                printf("%i ", homographies[lh_index][i][j]);
+                printf("%f ", (double)homographies[lh_index][i][j]);
             }
             printf("\n");
         }
